refactor: delete dead context.ts; relocate tool-result-truncation to pi (#632 segment 1, last cut)
Audit of the last two leaf files in segment 1:

- context.ts (pruneImages): exported but no production callers — only
  its own tests. Deleted entirely.
- tool-result-truncation.ts (truncateToolResultText): used by pi-mono
  to cap tool output before feeding back to the SDK. Pi-bound.
  Relocated to src/agent/runners/pi/.

Concludes segment 1 of #632 (leaf cleanup).

- src/legacy/core/context.{ts,test.ts}: deleted
- src/legacy/core/tool-result-truncation.{ts,test.ts} → src/agent/runners/pi/
- src/legacy/core/pi-mono.ts: import path updated

tsc clean; 1165/1165 tests pass (= 1168 minus 3 pruneImages tests).
