+++ src/agent/runners/pi/tool-result-truncation.test.ts
