+++ src/agent/runners/pi/tool-result-truncation.ts
